Extract AutoReReviewController from IssueTracker (CROW-1094)
Move the auto-re-request-review watcher (CROW-921) — applyAutoReRequestReview,
attemptReRequestReview, and the skip-reason / eligibility / round-key statics —
into its own collaborator, owning its in-flight / attempted / failure-count
bookkeeping. It reaches the shared codeBackend, steady-state logger,
agent-settled check, and the autoReRequestReview toggle (all widened to internal,
kept on the tracker) via an unowned back-reference. The fire-and-forget dispatch
Task uses [weak self] so it no-ops rather than touching a freed tracker — the
same pattern BoardPoller's auto-create dispatch already uses.

applyPRStatuses drives it; an extension IssueTracker re-exposes the
IssueTracker.<symbol> / tracker.<member> surface the tests use (skip-reason type,
statics, and the in-flight/attempted/failure-count proxies).

swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
